alt.bg to comp + fix scripture + rename to bgfileid
