+++ src/components/output/BaseOutput.vue
@@ -13,7 +13,28 @@ export default {
     },
     presentationType () {
       return presentationTypes.find(t => t.id === this.presentation?.type)
+    },
+    backgroundImageUrl () {
+      return this.$store.getMediaUrl(this.settings.bgfileId)
+    },
+    style () {
+      const style = {}
+
+      if (this.backgroundImageUrl) {
+        style.backgroundImage = `url(${this.backgroundImageUrl})`
+      }
+
+      return style
     }
   }
 }
 </script>
+
+<style>
+.bg-output-beamer {
+  width: 100%;
+  height: 100%;
+  background-size: cover;
+  background-position: center;
+}
+</style>
