+++ src/components/song/SongSettings.vue
@@ -38,19 +38,8 @@
           </div>
         </div>
       </q-tab-panel>
-
       <q-tab-panel name="background">
-        <q-file v-model="background" accept="image/*" label="Selecteer achtergrondafbeelding" outlined @update:model-value="updateBackground">
-          <template #prepend>
-            <q-icon name="image" />
-          </template>
-
-          <template v-if="settings.fileId" #append>
-            <q-icon name="cancel" class="cursor-pointer" @click="resetBackground" />
-          </template>
-        </q-file>
-
-        <img :src="backgroundUrl" class="q-mt-sm full-width">
+        <BackgroundSetting v-model:bgfileId="settings.bgfileId" />
       </q-tab-panel>
     </q-tab-panels>
   </div>
@@ -58,19 +47,15 @@
 
 <script>
 import BaseSettings from '../presentation/BaseSettings.vue'
+import BackgroundSetting from '../presentation/BackgroundSetting.vue'
 
 export default {
+  components: { BackgroundSetting },
   extends: BaseSettings,
   data () {
     return {
       tab: 'text',
-      isTranslating: false,
-      background: null
-    }
-  },
-  computed: {
-    backgroundUrl () {
-      return this.$store.getMediaUrl(this.settings.fileId || this.$store.service.backgroundImageId)
+      isTranslating: false
     }
   },
   methods: {
@@ -85,13 +70,6 @@ export default {
       } finally {
         this.isTranslating = false
       }
-    },
-    updateBackground (file) {
-      this.settings.fileId = this.$store.addMedia(file)
-    },
-    resetBackground () {
-      this.settings.fileId = null
-      this.background = null
     }
   }
 }
